Issue38 - Moved util classes to folders in the integration/unit folders. Fixed integration test to work with new getDb method
